pyproject.toml
[project]
name = "work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "azure-identity>=1.25.1",
    "python-dotenv>=1.0.0",
    "openai>=2.8.1",
    "streamlit>=1.28.0",
    "pyyaml>=6.0",
]
